Made a safety update to Random Playerbot manager
Found an existing ratio for handling faction loading, with my changes forcing all created characters to being one faction only, this ratio logic would break.

So I've implemented a fix to retain original logic if mine isn't used but to fix an issue that would cause a bug when my logic is enabled for either alliance or horde.